♻️ refactor: extract BG continuation into separate extension files
Splits the background continuation code out of SimulationViewModel.swift
and SimulationView.swift into dedicated +Background.swift files to stay
within SwiftLint file_length / type_body_length limits.

- SimulationViewModel+Background.swift: enable/disable, scenePhase
  handlers, and private GPU↔CPU switching helpers.
- SimulationView+Background.swift: the toggle button view helper.
- isBackgroundContinuationEnabled and errorMessage drop private(set)
  to allow cross-file extension mutation (same module only).
- handleEvent marked with explicit cyclomatic_complexity exception.

Progress in #84
